+++ changelog.md
@@ -6,6 +6,7 @@
 * Enabled minify
 + Check if mobile data is available on the device (targeted at tablets)
 + Ads in the play services and network tab
++ More information in the device tab including IMEI, root status and version name
 
 #1.20160724
 + Added main interface